refactor: move day indicator comp to shared folder
